+++ 9.txt
@@ -1,0 +1,10 @@
+Simon
+1000
+Robin
+1000
+AlanTuring
+10000
+https://youtu.be/dQw4w9WgXcQ
+10000
+NewScore
+10000
