+++ README.md
@@ -1,2 +1,4 @@
 # Demineur
 Short mine sweeper coded in C
+LANDRY Simon
+KHATIB Robin
